Create index.html in the root
I created index.html file in the root and deleted it from assets folder because it was giving error when i click on my site link. The problem was that the index.html file was in the subfolder. The correct approach is to save index.html at root out of any subfolder.